Addressing comments

+++ Pal/lib/crypto/adapters/mbedtls_dh.c
@@ -27,7 +27,7 @@
 #define BITS_PER_BYTE 8
 
 /* This is declared in pal_internal.h, but that can't be included here. */
-int _DkRandomBitsRead(void *buffer, int size);
+size_t _DkRandomBitsRead(void *buffer, int size);
 
 /* Wrapper to provide mbedtls the RNG interface it expects. It passes an
  * extra context parameter, and expects a return value of 0 for success
